+++ contracts/interfaces/IDeployer.sol
@@ -47,7 +47,7 @@ struct DeploymentParams {
     bool enableIssuancePremium; // whether to enable the issuance premium
     //
     // === BackingManager ===
-    uint48 tradingDelay; // {s} how long to wait until starting auctions after switching basket
+    uint48 tradingDelay; // {s} delay before automatic trading after a default-triggered basket switch
     uint48 batchAuctionLength; // {s} the length of a Gnosis EasyAuction
     uint48 dutchAuctionLength; // {s} the length of a falling-price dutch auction
     uint192 backingBuffer; // {1} how much extra backing collateral to keep